refactor: 씨앗 UI Presentation and Container 디자인 패턴으로 리팩토링
