+++ src/types/seeds.ts
@@ -1,0 +1,17 @@
+export type SeedHistoryRawItem = {
+  id: number;
+  userId: string;
+  amount: number;
+  type: string;
+  description: string;
+  balanceSnapshot: number;
+  createdAt: string;
+}
+
+export type SeedsResponse = {
+  content: SeedHistoryRawItem[];
+  page: number;
+  pageSize: number;
+  totalPages: number;
+  totalElements: number;
+}
